Refs #37: Add new packages to organize classes, add CyRest dependencies to the pom.xml
