+++ pom.xml
@@ -174,6 +174,20 @@
 			<version>2.0.2-beta</version>
 			<scope>test</scope>
 		</dependency>
+		
+		<!-- CyRest -->
+		<dependency>
+			<groupId>javax.ws.rs</groupId>
+			<artifactId>javax.ws.rs-api</artifactId>
+			<version>2.0</version>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
+			<groupId>io.swagger</groupId>
+			<artifactId>swagger-annotations</artifactId>
+			<version>1.5.7</version>
+			<scope>provided</scope>
+		</dependency>
 	</dependencies>
 
 	<groupId>com.dpgil.path_linker</groupId>
