+++ src/main/java/com/dpgil/pathlinker/path_linker/internal/event/PathLinkerNodeSelectionListener.java
@@ -1,9 +1,11 @@
-package com.dpgil.pathlinker.path_linker.internal;
+package com.dpgil.pathlinker.path_linker.internal.event;
 
 import org.cytoscape.model.CyNetwork;
 import org.cytoscape.model.events.RowSetRecord;
 import org.cytoscape.model.events.RowsSetEvent;
 import org.cytoscape.model.events.RowsSetListener;
+
+import com.dpgil.pathlinker.path_linker.internal.view.PathLinkerControlPanel;
 
 /** Listener class for _loadNodeToSourceButton and _loadNodeToTargetButton in PathLinkerPanel class */
 public class PathLinkerNodeSelectionListener implements RowsSetListener {
