+++ src/main/java/com/dpgil/pathlinker/path_linker/internal/view/PathLinkerAboutMenuDialog.java
@@ -1,4 +1,4 @@
-package com.dpgil.pathlinker.path_linker.internal;
+package com.dpgil.pathlinker.path_linker.internal.view;
 
 import java.awt.BorderLayout;
 import java.awt.Desktop;
