+++ src/main/java/com/dpgil/pathlinker/path_linker/internal/view/PathLinkerResultPanel.java
@@ -1,6 +1,4 @@
-package com.dpgil.pathlinker.path_linker.internal;
-
-import com.dpgil.pathlinker.path_linker.internal.Algorithms.Path;
+package com.dpgil.pathlinker.path_linker.internal.view;
 
 import java.awt.Component;
 import java.awt.Container;
@@ -38,6 +36,8 @@ import org.cytoscape.model.CyNetwork;
 import org.cytoscape.model.CyNetworkManager;
 import org.cytoscape.model.CyNode;
 import org.cytoscape.model.CyTableUtil;
+
+import com.dpgil.pathlinker.path_linker.internal.util.Algorithms.Path;
 
 /**
  * // -------------------------------------------------------------------------
